requirements.txt
psycopg2
plotly==4.14.3
pandas==1.2.3
numpy==1.20.2
configparser==5.0.2
jupyter-dash==0.4.0
dash-bootstrap-components==0.12.0
dash-html-components==1.1.3
boto3
PyGithub
uwsgi